requirements.txt
Django==1.10.0
Pillow==3.4.1
django-ckeditor==5.1.1
django-tagging==0.4.5
django-ordered-model==1.3.0
six==1.10.0 # required by django-ordered-model
mysqlclient==1.3.9
pytz
